# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Assistant API",
    "description": "AI assistant API — A2A protocol (Agent-to-Agent), chat, and workflow management.\n\n**Authentication**: protected endpoints require `Authorization: Bearer <token>`.\nThe token is set via `--auth-token` / `ASSISTANT_WEB_TOKEN` on the server.",
    "license": {
      "name": "MIT",
      "identifier": "MIT"
    },
    "version": "0.1.135"
  },
  "servers": [
    {
      "url": "/",
      "description": "Local assistant server"
    }
  ],
  "paths": {
...
    "/api/agents": {
      "get": {
        "tags": [
          "agents"
        ],
        "summary": "`GET /api/agents` — list all registered agents.",
        "operationId": "list_agents",
        "responses": {
          "200": {
            "description": "List of registered agents",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AgentSummary"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "security": [
          {
            "bearer_token": []
          }
        ]
      },
      "post": {
        "tags": [
          "agents"
        ],
        "summary": "`POST /api/agents` — register a new agent.",
        "operationId": "register_agent",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterAgentRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Agent registered",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentDetail"
                }
              }
            }
          },
          "400": {
            "description": "Invalid agent card"
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Internal server error"
          }
        },
        "security": [
          {
            "bearer_token": []
          }
        ]
      }
    },
    "/api/agents/{id}": {
      "get": {
        "tags": [
          "agents"
        ],
        "summary": "`GET /api/agents/{id}` — get an agent by ID.",
        "operationId": "get_agent",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Agent ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Agent detail",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentDetail"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Agent not found"
          }
        },
        "security": [
          {
            "bearer_token": []
          }
        ]
      },
      "put": {
        "tags": [
          "agents"
        ],
        "summary": "`PUT /api/agents/{id}` — update an agent's card.",
        "operationId": "update_agent",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Agent ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateAgentRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Updated agent detail",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentDetail"
                }
              }
            }
          },
          "400": {
            "description": "Invalid agent card"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Agent not found"
          }
        },
        "security": [
          {
            "bearer_token": []
          }
        ]
      },
      "delete": {
        "tags": [
          "agents"
        ],
        "summary": "`DELETE /api/agents/{id}` — remove a registered agent.",
        "operationId": "delete_agent",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Agent ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Deleted"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Agent not found"
          }
        },
        "security": [
          {
            "bearer_token": []
          }
        ]
      }
    },
    "/api/agents/{id}/set-default": {
      "post": {
        "tags": [
          "agents"
        ],
        "summary": "`POST /api/agents/{id}/set-default` — set an agent as the default.",
        "operationId": "set_default_agent",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Agent ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Updated agent detail",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentDetail"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Agent not found"
          }
        },
        "security": [
          {
            "bearer_token": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AgentDetail": {
        "type": "object",
        "description": "Full agent detail, including the serialized agent card.",
        "required": [
          "id",
          "is_default",
          "card"
        ],
        "properties": {
          "card": {
            "description": "Full AgentCard serialized as JSON."
          },
          "id": {
            "type": "string"
          },
          "is_default": {
            "type": "boolean"
          }
        }
      },
...
      "AgentSummary": {
        "type": "object",
        "description": "A brief summary of a registered agent.",
        "required": [
          "id",
          "name",
          "version",
          "description",
          "is_default",
          "skill_count",
          "interface_count"
        ],
        "properties": {
          "description": {
            "type": "string"
          },
          "id": {
            "type": "string"
          },
          "interface_count": {
            "type": "integer",
            "minimum": 0
          },
          "is_default": {
            "type": "boolean"
          },
          "name": {
            "type": "string"
          },
          "skill_count": {
            "type": "integer",
            "minimum": 0
          },
          "version": {
            "type": "string"
          }
        }
      },
...
      "RegisterAgentRequest": {
        "type": "object",
        "description": "Body for `POST /api/agents`.",
        "required": [
          "card"
        ],
        "properties": {
          "card": {
            "description": "Full AgentCard JSON."
          },
          "set_default": {
            "type": [
              "boolean",
              "null"
            ],
            "description": "If `true`, make this agent the default after registration."
          }
        }
      },
...
      "UpdateAgentRequest": {
        "type": "object",
        "description": "Body for `PUT /api/agents/{id}`.",
        "required": [
          "card"
        ],
        "properties": {
          "card": {
            "description": "Updated full AgentCard JSON."
          }
        }
      },
...
    },
    "securitySchemes": {
      "bearer_token": {
        "type": "http",
        "scheme": "bearer",
        "description": "Bearer token. Pass the token issued by the server via `Authorization: Bearer <token>`. Configure it with `--auth-token` / `ASSISTANT_WEB_TOKEN`."
      },
      "oauth2": {
        "type": "oauth2",
        "flows": {
          "authorizationCode": {
            "authorizationUrl": "/oauth/authorize",
            "tokenUrl": "/oauth/token",
            "refreshUrl": "/oauth/token",
            "scopes": {
              "api_keys:read": "List API keys",
              "api_keys:write": "Create and revoke API keys",
              "conversations:read": "Read conversations",
              "conversations:write": "Create and send messages",
              "org:manage": "Full organization management",
              "personas:read": "Read personas",
              "personas:write": "Create and update personas",
              "skills:read": "Read skills",
              "skills:write": "Create and update skills",
              "spaces:read": "Read spaces",
              "spaces:write": "Create and update spaces",
              "users:read": "Read users",
              "users:write": "Create and update users"
            }
          },
          "clientCredentials": {
            "tokenUrl": "/oauth/token",
            "scopes": {
              "api_keys:read": "List API keys",
              "api_keys:write": "Create and revoke API keys",
              "conversations:read": "Read conversations",
              "conversations:write": "Create and send messages",
              "org:manage": "Full organization management",
              "personas:read": "Read personas",
              "personas:write": "Create and update personas",
              "skills:read": "Read skills",
              "skills:write": "Create and update skills",
...
  },
...
}